refactor: relocation package directory
