+++ src/main/java/com/finance/finportfolio/domain/finance/domain/PortfolioRepository.java
@@ -1,4 +1,4 @@
-package com.finance.finportfolio.domain;
+package com.finance.finportfolio.domain.finance.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
