+++ src/main/java/com/finance/finportfolio/domain/post/domain/Post.java
@@ -1,4 +1,4 @@
-package com.finance.finportfolio.domain;
+package com.finance.finportfolio.domain.post.domain;
 
 import java.time.LocalDateTime;
