+++ src/main/java/com/finance/finportfolio/global/error/GlobalExceptionHandler.java
@@ -1,11 +1,9 @@
-package com.finance.finportfolio.global;
+package com.finance.finportfolio.global.error;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import com.finance.finportfolio.dto.ErrorResponse;
 
 import lombok.extern.slf4j.Slf4j;
