+++ src/test/java/com/finance/finportfolio/dto/PostSaveRequestDtoTest.java
@@ -5,7 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import com.finance.finportfolio.domain.Post;
+import com.finance.finportfolio.domain.post.domain.Post;
+import com.finance.finportfolio.domain.post.dto.PostSaveRequestDto;
 
 class PostSaveRequestDtoTest {
